+++ packages/shared/src/types/localization.ts
@@ -1153,7 +1153,6 @@ export type __internal_LocalizationResource = {
         domainLabel: LocalizationValue;
         signOnUrlLabel: LocalizationValue;
         issuerLabel: LocalizationValue;
-        certificateLabel: LocalizationValue;
         menuAction__edit: LocalizationValue;
         menuAction__activate: LocalizationValue;
         menuAction__deactivate: LocalizationValue;
@@ -1382,12 +1381,21 @@ export type __internal_LocalizationResource = {
         badge__verified: LocalizationValue;
         badge__unverified: LocalizationValue;
         verifiedAtLabel: LocalizationValue<'date'>;
+        removeButtonTooltip__lastVerifiedDomain: LocalizationValue;
+        removeButtonTooltip__lastVerifiedDomainActive: LocalizationValue;
         txtRecord: {
           instructions: LocalizationValue;
           typeLabel: LocalizationValue;
           hostLabel: LocalizationValue;
           valueLabel: LocalizationValue;
         };
+      };
+      removeDomainDialog: {
+        title: LocalizationValue;
+        subtitle__active: LocalizationValue<'domain'>;
+        subtitle__inactive: LocalizationValue<'domain'>;
+        cancelButton: LocalizationValue;
+        removeButton: LocalizationValue;
       };
     };
     testConfigurationStep: {
@@ -1835,7 +1843,6 @@ export type __internal_LocalizationResource = {
         title: LocalizationValue;
         ssoUrlLabel: LocalizationValue;
         issuerLabel: LocalizationValue;
-        certificateLabel: LocalizationValue;
         configureAgainLink: LocalizationValue;
       };
       resetSection: {